quotes/core/requirements.txt
Django==4.1.13
djangorestframework==3.15.1
numpy==1.26.4
pandas==2.2.2
scikit_learn==1.4.2
